refactor(api/app): flatten inngest dir to mirror api/platform
- api/app/src/inngest/client/client.ts -> client.ts (drop wrapper dir).
- api/app/src/inngest/workflow/infrastructure/record-activity.ts ->
  workflow/record-activity.ts (drop single-child category subdir).
- Fix client.ts relative imports for new location (../env, ./schemas/app).
- Update consumers in inngest/index.ts and lib/activity.ts.
